- Fixing Order Tracking screen behaviour - Fixed Hide Recommeded - Fixed App crash when clicking on track order - Fixed Item list getting cropped(restron product list) - Order status issue fixed - Change call now to 'call me' - From and To address should not be same fix - Fixed App crash while switching from orders to search - Rating screen blank button - Delivery At to be put at default address - App crash when clicking on track order fix - Guest login when tries to view the product it crashes fix - Your order is being placed text added - Delivery message needs to be generalised "Your order is ready and will soon be picked up by <Driver name>" done - Order not being cancelled even after clicking cancel button - Coming soon store when clicked should show - This store is coming soon - Prescription upload option is not working - Color code to be changed for View Card bottom section fix - (Minus) sign needs to be added where discount applied. - "Rejection screen :   - Refund message in order status tracking space   - Append 'Pay' before the value of cart ex : Pay $30   - User is allowed to review only once per order   - Replace cart pop up fix   - Prescription image shown   - increment on customized items is working   - View Cart colour code to be consistent in search tab fix   - Disabled store search   - Continous loader in order history screeen, if login as a guest fix   - Remove search button from saved addresses fix   - Remove payment method untill payment compliances fix   - App crash at profile and search during guest mode fix   - Store name to be populated with store name selected by the user fix   - Update large cards to maintain consistency fix

+++ app/src/main/java/com/fidoo/user/view/OrderDetailsActivity.kt
@@ -171,7 +171,7 @@ class OrderDetailsActivity : com.fidoo.user.utils.BaseActivity() {
             val deliveryChargeWithTax = mModelData.deliveryCharge + mModelData.tax
             delivery_charge.text = resources.getString(R.string.ruppee) + "" + deliveryChargeWithTax
             delivery_coupon_label.text = "Delivery Discount (" +mModelData.delivery_coupon_name +")"
-            delivery_coupon.text = "-" +resources.getString(R.string.ruppee) + "" + mModelData.delivery_discount
+            delivery_coupon.text = "-" +resources.getString(R.string.ruppee) + "" + mModelData.delivery_discount.toFloat()
             grand_price.text = resources.getString(R.string.ruppee) + "" + mModelData.totalPrice
             sub_total.text = resources.getString(R.string.ruppee) + "" + mModelData.totalPrice
         })
